+++ src/plugins/index_pattern_management/server/routes/resolve_index.ts
@@ -29,7 +29,7 @@
  */
 
 import { schema } from '@osd/config-schema';
-import { IRouter } from 'src/core/server';
+import { IRouter, LegacyAPICaller } from 'src/core/server';
 
 export function registerResolveIndexRoute(router: IRouter): void {
   router.get(
@@ -59,25 +59,16 @@ export function registerResolveIndexRoute(router: IRouter): void {
         : null;
 
       const dataSourceId = req.query.data_source;
-      if (dataSourceId) {
-        const result = await (
-          await context.dataSource.opensearch.getClient(dataSourceId)
-        ).indices.resolveIndex({
-          name: encodeURIComponent(req.params.query),
-          expand_wildcards: req.query.expand_wildcards,
-        });
-        return res.ok({ body: result.body });
-      }
+      const caller = dataSourceId
+        ? context.dataSource.opensearch.legacy.getClient(dataSourceId).callAPI
+        : context.core.opensearch.legacy.client.callAsCurrentUser;
 
-      const result = await context.core.opensearch.legacy.client.callAsCurrentUser(
-        'transport.request',
-        {
-          method: 'GET',
-          path: `/_resolve/index/${encodeURIComponent(req.params.query)}${
-            queryString ? '?' + new URLSearchParams(queryString).toString() : ''
-          }`,
-        }
-      );
+      const result = await caller('transport.request', {
+        method: 'GET',
+        path: `/_resolve/index/${encodeURIComponent(req.params.query)}${
+          queryString ? '?' + new URLSearchParams(queryString).toString() : ''
+        }`,
+      });
       return res.ok({ body: result });
     }
   );